dw-np-prod-promo-controls-us-es4-function/requirements.txt
functions-framework==3.*
requests
google-cloud-bigquery
google-cloud-storage
google-cloud-bigquery-storage
google-cloud-logging
google-auth
google-api-core